feat(tasks): unify glob strategy for tasks and dependencies
